requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
amqp==5.0.6
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==4.1.0
blessings==1.7
blobfile==1.2.5
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
clip @ git+https://github.com/openai/CLIP.git@fa56f2525191a013533338f137aab59ac36d8c26
cycler==0.10.0
d2l==0.17.0
DALL-E @ git+https://github.com/openai/DALL-E@5be4b236bc3ade6943662354117a0e83752cc322
debugpy==1.4.3
decorator==5.1.0
deep-translator==1.5.4
deepspeed @ file:///tmp/Deepspeed/dist/deepspeed-0.5.5%2B1fc74cb-cp38-cp38-linux_x86_64.whl
defusedxml==0.7.1
discord==1.7.3
discord.py @ git+https://github.com/Rapptz/discord.py.git@fbc4a51c355c029af61363038d0a8a9ac0bd8ace
efficientnet-pytorch==0.6.3
einops==0.3.2
entrypoints==0.3
filelock==3.0.12
fsspec==2021.7.0
ftfy==6.0.3
future==0.18.2
glide==0.0.0
-e git+git@github.com:luc-leonard/clip_generators.git@0d6cceefc0a7d7615a0f7f19e77d41495683095b#egg=glide_text2im&subdirectory=clip_generators/models/glide/glide-text2im
google-auth==1.35.0
google-auth-oauthlib==0.4.5
gpustat==0.6.0
grpcio==1.39.0
huggingface-hub==0.0.17
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
irc==19.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.logging==3.1.0
jaraco.stream==3.0.2
jaraco.text==3.5.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kombu==5.1.0
kornia==0.5.8
lmdb==1.2.1
lpips==0.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
mpi4py==3.1.1
multidict==5.1.0
munch==2.5.0
mypy==0.930
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
ninja==1.10.2.2
notebook==6.4.4
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
omegaconf==2.1.0
opencv-python==4.5.4.60
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==1.0.0
pretrainedmodels==0.7.4
progressbar2==3.53.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.11.0
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-utils==2.5.6
pytorch-lightning==1.4.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
rudalle==0.4.0
sacremoses==0.0.45
scipy==1.7.1
segmentation-models-pytorch==0.1.3
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3
taming-transformers==0.0.1
tempora==4.1.1
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
terminado==0.12.1
testpath==0.5.0
timm==0.3.4
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.0
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchmetrics==0.5.0
torchvision==0.11.2+cu113
tornado==6.1
tqdm==4.62.1
traitlets==5.1.0
transformers==4.10.2
triton==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
yarl==1.6.3
youtokentome==1.0.6
